Workaround the bug in geography_out which excludes bounding box info
_ST_Expand(geography, float) adds the expansion info to the in-memory
structure, but that information is lost when the value is converted into
a text representation. That later causes XL datanode to run with an
incorrect plan.

Fix this for now by moving _ST_Expand (which ST_DWitin uses) into ST_DWithin
itself, thus avoiding SQL function preprocessing and working around the
issue.